Plotly graph now displays data
The graph now displays data, consumes abot 160 mb of data on the web browser but slows down/is buggy right now. It cannot be the graph or the volume of data causing it since it runs fine on jupyter. I suspect it is the callback function being called constantly. Will need to either implement debouncer, flags or implement state machines for updating data.